move canonicalurl tests to new file
